refactor: 식습관 분석 코드 모듈화 진행(함수화 리팩토링 필요)
